+++ towerdiscord/hint.go
@@ -8,10 +8,6 @@ type MimetypeHint interface {
 	Mimetype() string
 }
 
-type LengthHint interface {
-	Len() int
-}
-
 type ImageSizeHint interface {
 	ImageSize() (width, height int)
 }